+++ NBodies/Physics/Structs/GridInfo.cs
@@ -8,18 +8,18 @@ namespace NBodies.Physics
 {
     public struct GridInfo
     {
-        public long OffsetX;
-        public long OffsetY;
-        public long MinX;
-        public long MinY;
-        public long MaxX;
-        public long MaxY;
-        public long Columns;
-        public long Rows;
-        public long Size;
-        public long IndexOffset;
+        public int OffsetX;
+        public int OffsetY;
+        public int MinX;
+        public int MinY;
+        public int MaxX;
+        public int MaxY;
+        public int Columns;
+        public int Rows;
+        public int Size;
+        public int IndexOffset;
 
-        public GridInfo(long offX, long offY, long idxOff, long minX, long minY, long maxX, long maxY, long cols, long rows)
+        public GridInfo(int offX, int offY, int idxOff, int minX, int minY, int maxX, int maxY, int cols, int rows)
         {
             OffsetX = offX;
             OffsetY = offY;
@@ -34,7 +34,7 @@ namespace NBodies.Physics
             Size = ((cols + 1) * (rows + 1));
         }
 
-        public void Set(long offX, long offY, long idxOff, long minX, long minY, long maxX, long maxY, long cols, long rows)
+        public void Set(int offX, int offY, int idxOff, int minX, int minY, int maxX, int maxY, int cols, int rows)
         {
             OffsetX = offX;
             OffsetY = offY;
